Fix multi-release jar to avoid API differences between versions.
This fixes the following warnings from `jar --validate`:
```
entry: META-INF/versions/15/com/jcraft/jsch/jce/SignatureEd448.class, contains a class with different api from earlier version
entry: META-INF/versions/15/com/jcraft/jsch/jce/SignatureEd25519.class, contains a class with different api from earlier version
entry: META-INF/versions/24/com/jcraft/jsch/jce/MLKEM768.class, contains a class with different api from earlier version
entry: META-INF/versions/24/com/jcraft/jsch/jce/MLKEM1024.class, contains a class with different api from earlier version
```
