+++ pom.xml
@@ -694,10 +694,12 @@
                         <exclude>com/jcraft/jsch/JplLogger.class</exclude>
                         <exclude>com/jcraft/jsch/UnixDomainSocketFactory.class</exclude>
                         <exclude>com/jcraft/jsch/jce/KeyPairGenEdDSA.class</exclude>
+                        <exclude>com/jcraft/jsch/jce/MLKEM.class</exclude>
                         <exclude>com/jcraft/jsch/jce/MLKEM1024.class</exclude>
                         <exclude>com/jcraft/jsch/jce/MLKEM768.class</exclude>
                         <exclude>com/jcraft/jsch/jce/SignatureEd25519.class</exclude>
                         <exclude>com/jcraft/jsch/jce/SignatureEd448.class</exclude>
+                        <exclude>com/jcraft/jsch/jce/SignatureEdDSA.class</exclude>
                         <exclude>com/jcraft/jsch/jce/XDH.class</exclude>
                         <exclude>META-INF/versions/9/com/jcraft/jsch/JavaVersion.class</exclude>
                         <exclude>META-INF/versions/10/com/jcraft/jsch/JavaVersion.class</exclude>
