+++ src/main/java/com/jcraft/jsch/jce/SignatureEdDSA.java
@@ -26,9 +26,39 @@
 
 package com.jcraft.jsch.jce;
 
-public class SignatureEd448 extends SignatureEdDSA {
+abstract class SignatureEdDSA implements com.jcraft.jsch.SignatureEdDSA {
 
-  public SignatureEd448() {
-    throw new UnsupportedOperationException("SignatureEd448 requires Java15+.");
+  SignatureEdDSA() {
+    throw new UnsupportedOperationException("SignatureEdDSA requires Java15+.");
+  }
+
+  @Override
+  public void init() throws Exception {
+    throw new UnsupportedOperationException("SignatureEdDSA requires Java15+.");
+  }
+
+  @Override
+  public void setPubKey(byte[] y_arr) throws Exception {
+    throw new UnsupportedOperationException("SignatureEdDSA requires Java15+.");
+  }
+
+  @Override
+  public void setPrvKey(byte[] bytes) throws Exception {
+    throw new UnsupportedOperationException("SignatureEdDSA requires Java15+.");
+  }
+
+  @Override
+  public byte[] sign() throws Exception {
+    throw new UnsupportedOperationException("SignatureEdDSA requires Java15+.");
+  }
+
+  @Override
+  public void update(byte[] foo) throws Exception {
+    throw new UnsupportedOperationException("SignatureEdDSA requires Java15+.");
+  }
+
+  @Override
+  public boolean verify(byte[] sig) throws Exception {
+    throw new UnsupportedOperationException("SignatureEdDSA requires Java15+.");
   }
 }
